fixed 2 more VxWorks compiler warning in z017_io.c, z127_io.c

+++ DRIVERS/MDIS_LL/Z017/TOOLS/Z127_IO/COM/z127_io.c
@@ -196,16 +196,16 @@ int main(int argc, char *argv[])
 				printf("port : 31 30 29 28 27 26 25 24 23 22 21 20 19 18 17 16\n");
 				printf("state: ");
 				for (n=31; n>=16; n--)
-					printf("%d  ", (val >> n) & 1);
+					printf("%ld  ", (val >> n) & 1);
 				printf("\n\n");
 				printf("port : 15 14 13 12 11 10 9  8  7  6  5  4  3  2  1  0\n");
 				printf("state: ");
 				for (n=15; n>=0; n--)
-					printf("%d  ", (val >> n) & 1);
+					printf("%ld  ", (val >> n) & 1);
 				printf("\n\n");
 			} else {
 				printf("input state port #%d: %d (read-value=0x%x)\n",
-							port, (val & portBit) ? 1 : 0, val);
+							(int)port, (val & portBit) ? 1 : 0, (int)val);
 			}
 			if (getLoop)
 				UOS_Delay(2000);
@@ -216,7 +216,7 @@ int main(int argc, char *argv[])
 	|  set output           |
 	+----------------------*/
 	if (set != (-1)) {
-		printf("set output state port #%s: %d\n", portStr, set);
+		printf("set output state port #%s: %d\n", portStr, (int)set);
 		if ((M_setstat(path, set ? Z17_SET_PORTS : Z17_CLR_PORTS, portBit)) < 0) {
 			PrintError("setstat Z17_SET_PORTS/Z17_CLR_PORTS");
 			goto abort;
@@ -256,7 +256,7 @@ int main(int argc, char *argv[])
 		if (port == (-1))
 			printf("toggle output state ports %s\n", portStr);
 		else
-			printf("toggle output state port #%d\n", portBit);
+			printf("toggle output state port #%d\n", (int)portBit);
 
 		if ((M_setstat(path, Z17_TOG_PORTS, portBit)) < 0) {
 			PrintError("setstat Z17_TOG_PORTS");
